modifique navbar, modularize login, agregue font-awesome para iconos,agregue un fondo NO definitivo
