+++ client/index.html
@@ -7,11 +7,13 @@
       type="image/svg+xml"
       href="https://americansportsplanet.com/wp-content/uploads/2022/08/ASP-Logo-Temp.png"
     />
+    <link rel="stylesheet" href="/node_modules/@fortawesome/fontawesome-free/css/all.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sport Planet</title>
   </head>
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.jsx"></script>
+    <script src="/node_modules/@fortawesome/fontawesome-free/js/all.js"></script>
   </body>
 </html>
